+++ src/lib.rs
@@ -80,7 +80,7 @@ mod shortcut {
 
     use crate::{generator::GlobalGenerator, Scru64Id};
 
-    const DELAY: time::Duration = time::Duration::from_millis(64);
+    pub const DELAY: time::Duration = time::Duration::from_millis(64);
 
     /// Generates a new SCRU64 ID object using the global generator.
     #[doc = concat!("\n\n", include_str!("generator/doc_global_gen.md"), "\n\n")]
@@ -138,24 +138,22 @@ mod shortcut {
     /// Generates a new SCRU64 ID object using the global generator.
     #[doc = concat!("\n\n", include_str!("generator/doc_global_gen.md"), "\n\n")]
     ///
-    /// This function usually returns a value immediately, but if not possible, it sleeps and waits
-    /// for the next timestamp tick using the asynchronous sleep function provided as the argument.
-    /// Typical asynchronous sleep functions include `tokio::time::sleep`, `smol::Timer::after`,
-    /// and `async_std::task::sleep`.
+    /// This function usually returns a value immediately, but if not possible, it waits for the
+    /// next timestamp tick using the runtime-dependent yield function provided as the argument.
     ///
     /// # Panics
     ///
     /// Panics if the global generator is not properly configured.
     #[cfg(any(feature = "unstable", feature = "tokio"))]
-    pub async fn new_with<F, T>(mut async_sleep_fn: impl FnMut(time::Duration) -> F) -> Scru64Id
+    pub async fn new_with<F>(mut yield_fn: impl FnMut() -> F) -> Scru64Id
     where
-        F: std::future::Future<Output = T>,
+        F: std::future::Future,
     {
         loop {
             if let Some(value) = GlobalGenerator.generate() {
                 break value;
             } else {
-                async_sleep_fn(DELAY).await;
+                yield_fn().await;
             }
         }
     }
@@ -164,26 +162,22 @@ mod shortcut {
     /// global generator.
     #[doc = concat!("\n\n", include_str!("generator/doc_global_gen.md"), "\n\n")]
     ///
-    /// This function usually returns a value immediately, but if not possible, it sleeps and waits
-    /// for the next timestamp tick using the asynchronous sleep function provided as the argument.
-    /// Typical asynchronous sleep functions include `tokio::time::sleep`, `smol::Timer::after`,
-    /// and `async_std::task::sleep`.
+    /// This function usually returns a value immediately, but if not possible, it waits for the
+    /// next timestamp tick using the runtime-dependent yield function provided as the argument.
     ///
     /// # Panics
     ///
     /// Panics if the global generator is not properly configured.
     #[cfg(any(feature = "unstable", feature = "tokio"))]
-    pub async fn new_string_with<F, T>(
-        mut async_sleep_fn: impl FnMut(time::Duration) -> F,
-    ) -> String
+    pub async fn new_string_with<F>(mut yield_fn: impl FnMut() -> F) -> String
     where
-        F: std::future::Future<Output = T>,
+        F: std::future::Future,
     {
         loop {
             if let Some(value) = GlobalGenerator.generate() {
                 break value.into();
             } else {
-                async_sleep_fn(DELAY).await;
+                yield_fn().await;
             }
         }
     }
@@ -205,7 +199,7 @@ pub mod tokio {
     ///
     /// Panics if the global generator is not properly configured.
     pub async fn new() -> Scru64Id {
-        shortcut::new_with(tokio::time::sleep).await
+        shortcut::new_with(|| tokio::time::sleep(shortcut::DELAY)).await
     }
 
     /// Generates a new SCRU64 ID encoded in the 12-digit canonical string representation using the
@@ -219,7 +213,7 @@ pub mod tokio {
     ///
     /// Panics if the global generator is not properly configured.
     pub async fn new_string() -> String {
-        shortcut::new_string_with(tokio::time::sleep).await
+        shortcut::new_string_with(|| tokio::time::sleep(shortcut::DELAY)).await
     }
 
     /// Generates 100k monotonically increasing IDs.